Descarga del grafo como SVG
